moved base functionality of widgets to widget_base_class.py
