+++ customtkinter/customtkinter_draw_engine.py
@@ -2,7 +2,7 @@ import sys
 import tkinter
 from typing import Union
 
-from .widgets.customtkinter_canvas import CTkCanvas
+from .widgets.ctk_canvas import CTkCanvas
 
 
 class CTkDrawEngine:
