+++ customtkinter/widgets/ctk_label.py
@@ -1,18 +1,14 @@
 import tkinter
-import tkinter.ttk as ttk
 
-from .customtkinter_tk import CTk
-from .customtkinter_frame import CTkFrame
-from .customtkinter_canvas import CTkCanvas
-from ..appearance_mode_tracker import AppearanceModeTracker
+from .ctk_canvas import CTkCanvas
 from ..customtkinter_theme_manager import CTkThemeManager
 from ..customtkinter_settings import CTkSettings
 from ..customtkinter_draw_engine import CTkDrawEngine
+from .widget_base_class import CTkBaseClass
 
 
-class CTkLabel(tkinter.Frame):
+class CTkLabel(CTkBaseClass):
     def __init__(self, *args,
-                 master=None,
                  bg_color=None,
                  fg_color="default_theme",
                  text_color="default_theme",
@@ -22,54 +18,28 @@ class CTkLabel(tkinter.Frame):
                  text="CTkLabel",
                  text_font="default_theme",
                  **kwargs):
-        if master is None:
-            super().__init__(*args)
+
+        # transfer basic functionality (bg_color, size, appearance_mode, scaling) to CTkBaseClass
+        if "master" in kwargs:
+            super().__init__(*args, bg_color=bg_color, width=width, height=height, master=kwargs["master"])
+            del kwargs["master"]
         else:
-            super().__init__(*args, master=master)
+            super().__init__(*args, bg_color=bg_color, width=width, height=height)
 
-        # overwrite configure methods of master when master is tkinter widget, so that bg changes get applied on child CTk widget too
-        if isinstance(self.master, (tkinter.Tk, tkinter.Frame)) and not isinstance(self.master, (CTk, CTkFrame)):
-            master_old_configure = self.master.config
-
-            def new_configure(*args, **kwargs):
-                if "bg" in kwargs:
-                    self.configure(bg_color=kwargs["bg"])
-                elif "background" in kwargs:
-                    self.configure(bg_color=kwargs["background"])
-
-                # args[0] is dict when attribute gets changed by widget[<attribut>] syntax
-                elif len(args) > 0 and type(args[0]) == dict:
-                    if "bg" in args[0]:
-                        self.configure(bg_color=args[0]["bg"])
-                    elif "background" in args[0]:
-                        self.configure(bg_color=args[0]["background"])
-                master_old_configure(*args, **kwargs)
-
-            self.master.config = new_configure
-            self.master.configure = new_configure
-
-        # add set_appearance_mode method to callback list of AppearanceModeTracker for appearance mode changes
-        AppearanceModeTracker.add(self.change_appearance_mode, self)
-        self.appearance_mode = AppearanceModeTracker.get_mode()  # 0: "Light" 1: "Dark"
-
-        self.bg_color = self.detect_color_of_master() if bg_color is None else bg_color
+        # color
         self.fg_color = CTkThemeManager.theme["color"]["label"] if fg_color == "default_theme" else fg_color
         if self.fg_color is None:
             self.fg_color = self.bg_color
         self.text_color = CTkThemeManager.theme["color"]["text"] if text_color == "default_theme" else text_color
 
-        self.width = width
-        self.height = height
+        # shape
         self.corner_radius = CTkThemeManager.theme["shape"]["label_corner_radius"] if corner_radius == "default_theme" else corner_radius
 
-        if self.corner_radius * 2 > self.height:
-            self.corner_radius = self.height / 2
-        elif self.corner_radius * 2 > self.width:
-            self.corner_radius = self.width / 2
-
+        # text
         self.text = text
         self.text_font = (CTkThemeManager.theme["text"]["font"], CTkThemeManager.theme["text"]["size"]) if text_font == "default_theme" else text_font
 
+        # configure grid system (1x1)
         self.grid_rowconfigure(0, weight=1)
         self.grid_columnconfigure(0, weight=1)
 
@@ -78,6 +48,7 @@ class CTkLabel(tkinter.Frame):
                                 width=self.width,
                                 height=self.height)
         self.canvas.grid(row=0, column=0, sticky="nswe")
+        self.draw_engine = CTkDrawEngine(self.canvas, CTkSettings.preferred_drawing_method)
 
         self.text_label = tkinter.Label(master=self,
                                         highlightthickness=0,
@@ -87,46 +58,10 @@ class CTkLabel(tkinter.Frame):
                                         **kwargs)
         self.text_label.grid(row=0, column=0, padx=self.corner_radius)
 
-        self.draw_engine = CTkDrawEngine(self.canvas, CTkSettings.preferred_drawing_method)
-
-        super().configure(width=self.width, height=self.height)
-
         self.bind('<Configure>', self.update_dimensions)
         self.draw()
 
-    def destroy(self):
-        AppearanceModeTracker.remove(self.change_appearance_mode)
-        super().destroy()
-
-    def detect_color_of_master(self):
-        """ detect color of self.master widget to set correct bg_color """
-
-        if isinstance(self.master, CTkFrame):  # master is CTkFrame
-            return self.master.fg_color
-
-        elif isinstance(self.master, (ttk.Frame, ttk.LabelFrame, ttk.Notebook)):  # master is ttk widget
-            try:
-                ttk_style = ttk.Style()
-                return ttk_style.lookup(self.master.winfo_class(), 'background')
-            except Exception:
-                return "#FFFFFF", "#000000"
-
-        else:  # master is normal tkinter widget
-            try:
-                return self.master.cget("bg")  # try to get bg color by .cget() method
-            except Exception:
-                return "#FFFFFF", "#000000"
-
-    def update_dimensions(self, event):
-        # only redraw if dimensions changed (for performance)
-        if self.width != event.width or self.height != event.height:
-            self.width = event.width
-            self.height = event.height
-
-            # self.canvas.config(width=self.width, height=self.height)
-            self.draw()
-
-    def draw(self):
+    def draw(self, no_color_updates=False):
         requires_recoloring = self.draw_engine.draw_rounded_rect_with_border(self.width, self.height, self.corner_radius, 0)
 
         self.canvas.configure(bg=CTkThemeManager.single_color(self.bg_color, self.appearance_mode))
@@ -145,9 +80,6 @@ class CTkLabel(tkinter.Frame):
 
             self.text_label.configure(fg=CTkThemeManager.single_color(self.text_color, self.appearance_mode),
                                       bg=CTkThemeManager.single_color(self.bg_color, self.appearance_mode))
-
-    def config(self, *args, **kwargs):
-        self.configure(*args, **kwargs)
 
     def configure(self, *args, **kwargs):
         require_redraw = False  # some attribute changes require a call of self.draw() at the end
@@ -182,16 +114,3 @@ class CTkLabel(tkinter.Frame):
     def set_text(self, text):
         self.text = text
         self.text_label.configure(text=self.text, width=len(self.text))
-
-    def change_appearance_mode(self, mode_string):
-        if mode_string.lower() == "dark":
-            self.appearance_mode = 1
-        elif mode_string.lower() == "light":
-            self.appearance_mode = 0
-
-        if isinstance(self.master, (CTkFrame, CTk)):
-            self.bg_color = self.master.fg_color
-        else:
-            self.bg_color = self.master.cget("bg")
-
-        self.draw()
